reward for reporting a case is updated with fixedReward from Detective class

+++ Android/app/src/main/java/msp/powerrangers/logic/Confirmer.java
@@ -3,7 +3,7 @@ package msp.powerrangers.logic;
 
 public class Confirmer extends RoleForwarding {
 
-    int fixedReward = 2;
+    long fixedReward = 2;
     String caseId;
 
     public Confirmer(User user, String caseId) {
@@ -11,7 +11,7 @@ public class Confirmer extends RoleForwarding {
         this.user = user;
     }
 
-    public int getFixedReward(){
+    public long getFixedReward(){
         return fixedReward;
     }
